application/entity-resolution/pyproject.toml
[project]
name = "entity-resolution"
version = "0.1.0"
description = "Splink entity resolution on PostgreSQL"
requires-python = ">=3.10,<3.11"
dependencies = [
    "splink>=3.9.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "duckdb>=0.9.0",
    "sqlglot>=20.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["entity_resolution"]